added view page link to landing page generator
