+++ modules/flok/src/components/retreat-website/AddPageForm.tsx
@@ -38,7 +38,6 @@ function AddPageForm(props: AddPageFormProps) {
         )
       )
     }
-    console.log(result)
   }
   let formik = useFormik({
     initialValues: {
